+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-12461 Changed Open Data to Public Data naming
 * VFS-12396 Added cross-browser support for login message line-break formatting
 * VFS-12328 Added support for test custom login page
 * VFS-12528 Changed basic metadata to xattrs on harvester's indices page